# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BRAVE Program Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects for current team
      parameters:
        - name: teamId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive]
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
  /order-book-entries:
    get:
      operationId: listOrderBookEntries
      tags: [financials]
      summary: List order book entries
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: teamId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [verified]
      responses:
        "200":
          description: List of order book entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderBookEntry"
    post:
      operationId: createOrderBookEntry
      tags: [financials]
      summary: Create an order book entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBookEntryBody"
      responses:
        "201":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderBookEntry"
  /order-book-entries/{id}:
    get:
      operationId: getOrderBookEntry
      tags: [financials]
      summary: Get an order book entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order book entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderBookEntry"
    patch:
      operationId: updateOrderBookEntry
      tags: [financials]
      summary: Update an order book entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderBookEntryBody"
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderBookEntry"
    delete:
      operationId: deleteOrderBookEntry
      tags: [financials]
      summary: Delete an order book entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Revenue Entries
  /revenue-entries:
    get:
      operationId: listRevenueEntries
      tags: [financials]
      summary: List revenue entries
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: teamId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, submitted, verified, rejected]
      responses:
        "200":
          description: List of revenue entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueEntry"
    post:
      operationId: createRevenueEntry
      tags: [financials]
      summary: Create a revenue entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRevenueEntryBody"
      responses:
        "201":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueEntry"
  /revenue-entries/{id}:
    get:
      operationId: getRevenueEntry
      tags: [financials]
      summary: Get a revenue entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Revenue entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueEntry"
    patch:
      operationId: updateRevenueEntry
      tags: [financials]
      summary: Update a revenue entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRevenueEntryBody"
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueEntry"
  /revenue-entries/{id}/submit:
    post:
      operationId: submitRevenueEntry
      tags: [financials]
      summary: Submit a revenue entry for review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Entry submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueEntry"
  /revenue-entries/{id}/verify:
    post:
      operationId: verifyRevenueEntry
      tags: [financials]
      summary: Verify a revenue entry (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEntryBody"
      responses:
        "200":
          description: Entry verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueEntry"
  /revenue-entries/{id}/reject:
    post:
      operationId: rejectRevenueEntry
      tags: [financials]
      summary: Reject a revenue entry (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectEntryBody"
      responses:
        "200":
          description: Entry rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueEntry"

  # Admin review queue
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
        teamId:
          type: integer
        teamName:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, inactive]
        verifiedOrderBook:
          type: integer
        verifiedRevenue:
          type: integer
        clientCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - teamId
        - teamName
        - title
        - description
        - status
        - verifiedOrderBook
        - verifiedRevenue
        - clientCount
        - createdAt
        - updatedAt
...
    CreateProjectBody:
      type: object
      properties:
        title:
          type: string
          maxLength: 80
        description:
          type: string
          maxLength: 500
        teamId:
          type: integer
      required:
        - title
        - description
...
    OrderBookEntry:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        projectTitle:
          type: string
        teamId:
          type: integer
        teamName:
          type: string
        campusName:
          type: string
        clientName:
          type: string
        amount:
          type: integer
        verifiedAmount:
          type: ["integer", "null"]
        status:
          type: string
          enum: [verified]
        supportingDocUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        adminNotes:
          type: ["string", "null"]
        enteredBy:
          type: string
          enum: [student, admin]
        submittedAt:
          type: ["string", "null"]
          format: date-time
        verifiedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - projectTitle
        - teamId
        - teamName
        - campusName
        - clientName
        - amount
        - status
        - enteredBy
        - createdAt
    CreateOrderBookEntryBody:
      type: object
      properties:
        projectId:
          type: integer
        clientName:
          type: string
          maxLength: 100
        amount:
          type: integer
        supportingDocUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - projectId
        - clientName
        - amount
    UpdateOrderBookEntryBody:
      type: object
      properties:
        clientName:
          type: string
        amount:
          type: integer
        supportingDocUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    RevenueEntry:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        projectTitle:
          type: string
        teamId:
          type: integer
        teamName:
          type: string
        campusName:
          type: string
        clientName:
          type: string
        amount:
          type: integer
        verifiedAmount:
          type: ["integer", "null"]
        paymentDate:
          type: string
          format: date
        status:
          type: string
          enum: [draft, submitted, verified, rejected]
        paymentProofUrl:
          type: ["string", "null"]
        invoiceUrl:
          type: ["string", "null"]
        brdUrl:
          type: ["string", "null"]
        testimonialUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        adminNotes:
          type: ["string", "null"]
        enteredBy:
          type: string
          enum: [student, admin]
        submittedAt:
          type: ["string", "null"]
          format: date-time
        verifiedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - projectTitle
        - teamId
        - teamName
        - campusName
        - clientName
        - amount
        - paymentDate
        - status
        - enteredBy
        - createdAt
    CreateRevenueEntryBody:
      type: object
      properties:
        projectId:
          type: integer
        clientName:
          type: string
          maxLength: 100
        amount:
          type: integer
        paymentDate:
          type: string
          format: date
        paymentProofUrl:
          type: ["string", "null"]
        invoiceUrl:
          type: ["string", "null"]
        brdUrl:
          type: ["string", "null"]
        testimonialUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - projectId
        - clientName
        - amount
        - paymentDate
    UpdateRevenueEntryBody:
      type: object
      properties:
        clientName:
          type: string
        amount:
          type: integer
        paymentDate:
          type: string
          format: date
        paymentProofUrl:
          type: ["string", "null"]
        invoiceUrl:
          type: ["string", "null"]
        brdUrl:
          type: ["string", "null"]
        testimonialUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    VerifyEntryBody:
      type: object
      properties:
        verifiedAmount:
          type: integer
        adminNotes:
          type: ["string", "null"]
      required:
        - verifiedAmount
    RejectEntryBody:
      type: object
      properties:
        adminNotes:
          type: string
      required:
        - adminNotes
...